// src/interface/types.ts
export enum USER_ROLES {
  NORMAL = "NORMAL",
  ADMIN = "ADMIN",
}

export interface TokenPayload {
  id: string
  name: string
  role: USER_ROLES
}

export interface UserDB {
  id: string
  name: string
  email: string
  password: string
  role: USER_ROLES
  created_at: string
}

export interface UserModel {
  id: string
  name: string
  email: string
  password: string
  role: USER_ROLES
  createdAt: string
}